+++ tests_/conftest.py
@@ -1,6 +1,22 @@
 from pytest import fixture
 
+from data_safe_haven.config.context_settings import Context
 from data_safe_haven.external import AzureApi
+
+
+@fixture
+def context_dict():
+    return {
+        "admin_group_id": "d5c5c439-1115-4cb6-ab50-b8e547b6c8dd",
+        "location": "uksouth",
+        "name": "Acme Deployment",
+        "subscription_name": "Data Safe Haven (Acme)",
+    }
+
+
+@fixture
+def context(context_dict):
+    return Context(**context_dict)
 
 
 @fixture
@@ -50,7 +66,25 @@ def config_file(config_yaml, tmp_path):
 
 
 @fixture
-def mock_upload_blob(monkeypatch):
+def mock_download_blob(monkeypatch, context, config_yaml):
+    def mock_download_blob(
+        self,  # noqa: ARG001
+        blob_name: str,
+        resource_group_name: str,
+        storage_account_name: str,
+        storage_container_name: str,
+    ):
+        assert blob_name == context.config_filename
+        assert resource_group_name == context.resource_group_name
+        assert storage_account_name == context.storage_account_name
+        assert storage_container_name == context.storage_container_name
+        return config_yaml
+
+    monkeypatch.setattr(AzureApi, "download_blob", mock_download_blob)
+
+
+@fixture
+def mock_upload_blob(monkeypatch, context):
     def mock_upload_blob(
         self,  # noqa: ARG001
         blob_data: bytes | str,  # noqa: ARG001
@@ -59,6 +93,10 @@ def mock_upload_blob(monkeypatch):
         storage_account_name: str,  # noqa: ARG001
         storage_container_name: str,  # noqa: ARG001
     ):
+        assert blob_name == context.config_filename
+        assert resource_group_name == context.resource_group_name
+        assert storage_account_name == context.storage_account_name
+        assert storage_container_name == context.storage_container_name
         pass
 
     monkeypatch.setattr(AzureApi, "upload_blob", mock_upload_blob)
